parece que esta funcionando a transmissao

+++ src/server.c
@@ -6,80 +6,90 @@
 /*   By: mde-souz <mde-souz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/16 19:36:21 by mde-souz          #+#    #+#             */
-/*   Updated: 2024/07/20 18:02:14 by mde-souz         ###   ########.fr       */
+/*   Updated: 2024/07/22 22:01:44 by mde-souz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mini_talk.h"
 
 char	*byte;
-int	count;
 
-void	handler_sigusr(int sig)
+static void	handler_sigusr(int sig)
 {
 	if (sig == SIGUSR1)
 		*byte = '0';
 	else
 		*byte = '1';
 }
-void	get_bytes_print_string(int pid_client)
+static void	get_bytes_print_string(int pid_client)
 {
 	char	*byte_start;
 	int		i;
 	
-	count = 0;
 	byte_start = byte;
 	while (1)
 	{		
 		i = 8;
+		while (i)
+		{
+			if (i != 8)
+				while (kill(pid_client, SIGUSR1) == -1)
+					;
+			pause();
+			usleep(100);
+			byte++;
+			i--;
+		}
+		byte = byte_start;
+		if (ft_byte_to_char(byte_start) == 0)
+			break;
+		ft_printf(1, "%c", ft_byte_to_char(byte_start));
+		while (kill(pid_client, SIGUSR1) == -1)
+			;
+	}	
+}
+
+static int	get_pid_client(void)
+{
+	char	*byte_start;
+	int		pid_client;
+	int		i;
+	
+	byte_start = byte;
+	pid_client = 0;
+	while (1)
+	{
+		i = 8;
 		while (i--)
 		{
 			pause();
-			usleep(250);
-			//ft_printf(1, "count = %d\n",++count);
-			while (kill(pid_client, SIGUSR1) == -1)
-				;
 			byte++;
 		}
-		ft_printf(1, "%c", ft_byte_to_char(byte_start));
 		byte = byte_start;
-	}	
+		if (ft_byte_to_char(byte) == 0)
+			return (pid_client);
+		pid_client = pid_client * 10 + (ft_byte_to_char(byte) - '0');
+	}
 }
 
 int main(void)
 {
-	
-	char	*byte_start;
 	int		pid_client;
-	int		i;
 
 	signal(SIGUSR1,handler_sigusr);
 	signal(SIGUSR2,handler_sigusr);
 	byte = ft_calloc(9, sizeof(char));
 	if (!byte)
 		exit(EXIT_FAILURE);
-	ft_printf(1, "O PID É: %d",getpid());
-	byte_start = byte;
+	ft_printf(1, "SERVER PID: %d\n",getpid());
 	while (1)
 	{
-		ft_printf(1, "Primeiro programa sendo executado\n");
-		pid_client = 0;
-		while (1)
-		{
-			i = 8;
-			while (i--)
-			{
-				pause();
-				byte++;
-			}
-			byte = byte_start;
-			ft_printf(1, "digito pid %d\n", (ft_byte_to_char(byte) - '0'));
-			if (ft_byte_to_char(byte) == 0)
-				break;
-			pid_client = pid_client * 10 + (ft_byte_to_char(byte) - '0');
-		}
-		ft_printf(1, "Pid do programa sendo executado %d\n", pid_client);
-		get_bytes_print_string(pid_client);	
+		pid_client = get_pid_client();
+		get_bytes_print_string(pid_client);
+		while (kill(pid_client, SIGUSR1) == -1)
+			;
+		ft_printf(1,"\n");
 	}
+	free(byte);
 	return 0;
 }
